scMut/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scmut"
version = "0.1.0"
description = "A Python package for single-cell Mutation Parser in single-cell lineage tracing."
readme = "README.md"
requires-python = ">=3.7"
license = { text = "MIT" }
authors = [
  { name = "Kaiyu Wang" }
]
maintainers = [
  { name = "Kaiyu Wang", email = "wangkaiyu2020@sibcb.ac.cn" }
]
keywords = ["single-cell", "lineage tracing", "mutation analysis", "deep-learning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
]

dependencies = [
    "numpy",
    "pandas",
    "scipy",
    "scikit-learn",
    "umap-learn",
    "matplotlib",
    "seaborn",
    "tqdm",
    "anndata",
    "torch"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["scMut*"]
